requirements.txt
attrs==22.1.0
click==8.1.3
coverage==6.5.0
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.41
alembic==1.12.1
pydantic==2.5.0
pytest==8.4.0
PyYAML==6.0.1
python-multipart==0.0.6
requests==2.31.0
pulsar-client==3.6.1